# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      points:
        type: integer
      website:
        type: string
    type: object
  handler.PageResponse:
    properties:
      content:
        items:
          $ref: '#/definitions/handler.UserResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  handler.PublishRequest:
    properties:
      content:
        type: string
      type:
        type: string
    type: object
  handler.TransferRequest:
    properties:
      amount:
        type: integer
      from_id:
        type: integer
      to_id:
        type: integer
    type: object
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      points:
        type: integer
      website:
        type: string
    type: object
  handler.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      points:
        type: integer
      updated_at:
        type: string
      website:
        type: string
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      points:
        type: integer
      updatedAt:
        type: string
      website:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: High-performance Go API following Clean Architecture and DDD.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Demo Enterprise API
  version: "1.0"
paths:
  /api/notifications:
    post:
      consumes:
      - application/json
      description: send a notification (SQS or Memory)
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/handler.PublishRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send notification
      tags:
      - notifications
  /api/users:
    get:
      description: get all users with pagination support
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PageResponse'
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create a new user from JSON body
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /api/users/{id}:
    delete:
      description: delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - users
    get:
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: update user by ID from JSON body
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user
      tags:
      - users
  /api/users/sync/{externalID}:
    post:
      description: fetch user from external API and save it locally
      parameters:
      - description: External User ID
        in: path
        name: externalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync external user
      tags:
      - users
  /api/users/transfer:
    post:
      consumes:
      - application/json
      description: move points from one user to another atomically
      parameters:
      - description: Transfer details
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/handler.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Transfer points
      tags:
      - users
swagger: "2.0"